+++ ios/RNZcash.swift
@@ -384,6 +384,14 @@ class RNZcash: RCTEventEmitter {
                 wallet.subscribe()
                 let txs = try await wallet.synchronizer.allTransactions()
                 wallet.emitTxs(transactions: txs)
+                let balances = try await wallet.synchronizer.getAccountsBalances()
+                if let accountUUID = wallet.accountUUID,
+                  let accountBalance = balances[accountUUID]
+                {
+                  let data = wallet.createBalanceEventData(from: accountBalance)
+                  wallet.emit("BalanceEvent", data)
+                }
+
                 resolve(nil)
               case .failure:
                 reject("RescanError", "Failed to rescan wallet", genericError)
@@ -595,6 +603,8 @@ class WalletSynchronizer: NSObject {
   }
 
   func updateProcessorState(event: SynchronizerState) {
+    updateBalanceState(event: event)
+
     var scanProgress = 0
 
     switch event.internalSyncStatus {
@@ -621,7 +631,6 @@ class WalletSynchronizer: NSObject {
       "networkBlockHeight": self.processorState.networkBlockHeight,
     ]
     emit("UpdateEvent", data)
-    updateBalanceState(event: event)
   }
 
   func initializeProcessorState() {
@@ -631,16 +640,7 @@ class WalletSynchronizer: NSObject {
     )
   }
 
-  func updateBalanceState(event: SynchronizerState) {
-    guard let accountUUID = self.accountUUID else {
-      return
-    }
-
-    // Safely check if the account exists in the balances dictionary
-    guard let accountBalance = event.accountsBalances[accountUUID] else {
-      return
-    }
-
+  func createBalanceEventData(from accountBalance: AccountBalance) -> NSDictionary {
     // Account exists, safely access the balance properties
     let transparentBalance = accountBalance.unshielded
     let shieldedBalance = accountBalance.saplingBalance
@@ -655,7 +655,7 @@ class WalletSynchronizer: NSObject {
     let orchardAvailableZatoshi = orchardBalance.spendableValue
     let orchardTotalZatoshi = orchardBalance.total()
 
-    let data: NSDictionary = [
+    return [
       "alias": self.alias,
       "transparentAvailableZatoshi": String(transparentAvailableZatoshi.amount),
       "transparentTotalZatoshi": String(transparentTotalZatoshi.amount),
@@ -663,7 +663,20 @@ class WalletSynchronizer: NSObject {
       "saplingTotalZatoshi": String(saplingTotalZatoshi.amount),
       "orchardAvailableZatoshi": String(orchardAvailableZatoshi.amount),
       "orchardTotalZatoshi": String(orchardTotalZatoshi.amount),
-    ]
+    ] as NSDictionary
+  }
+
+  func updateBalanceState(event: SynchronizerState) {
+    guard let accountUUID = self.accountUUID else {
+      return
+    }
+
+    // Safely check if the account exists in the balances dictionary
+    guard let accountBalance = event.accountsBalances[accountUUID] else {
+      return
+    }
+
+    let data = createBalanceEventData(from: accountBalance)
     emit("BalanceEvent", data)
   }