+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- changed: iOS - Emit balance after rescan
 - fixed: iOS - get UUID after starting synchronizer
 
 ## 0.10.0 (2025-11-18)